requirements.txt
streamlit
google-generativeai
python-dotenv
groq
plotly
numpy<2